refactor(wix-vibe-headless): split AGENTS.md pinning out of deploy.cjs
`deploy.cjs` mixed two concerns: copying a vertical's files into `src/` AND
appending the project note to `AGENTS.md`. That note is vertical-agnostic and
idempotent, so a per-vertical deploy was an odd home for it — only the FIRST
vertical's deploy actually pinned (it returned `agentsMdPinned: true`); every
later vertical deploy silently skipped the append via the marker check.

Move the note into a dedicated `install/pin-agents-md.cjs` helper and have
`base44.md` STEP 1 call it explicitly, right after `deploy.cjs <vertical>`.
`deploy.cjs` is now purely file-copy; the AGENTS.md note is a single, clearly
separate step. Behaviour is unchanged (same idempotent note, still pinned once).
