+++ skills/wix-vibe-headless/install/deploy.cjs
@@ -9,7 +9,7 @@
 // No vertical arg -> deploys just the shared transport; re-run with the vertical once known.
 // NOTE: .cjs on purpose — the app is an ESM package ("type":"module"); a .js here would load as ESM
 // and require()/module.exports would throw.
-const { existsSync, cpSync, readFileSync, appendFileSync } = require('fs');
+const { existsSync, cpSync } = require('fs');
 
 const REF = '/app/.agents/skills/wix-vibe-headless/references';
 const VERTICALS = ['storefront', 'bookings', 'blog', 'cms', 'portfolio', 'pricing-plans', 'events', 'members'];
@@ -33,22 +33,5 @@ if (vertical && VERTICALS.includes(vertical) && existsSync(`${REF}/${vertical}/a
 } else {
   deployed.note = 'no vertical given — deployed the shared transport only; re-run: node deploy.cjs <vertical>';
 }
-
-// Pin the skill location + project facts into AGENTS.md so later turns (after this doc leaves
-// context) still know the rules. Idempotent.
-const NOTE = `
-
-## This app — a Wix-managed headless frontend (built with the Wix skills)
-
-This project is the **frontend for a Wix-managed business** — a REST client that talks directly to a live Wix site over \`WIX_CLIENT_ID\`. The **Wix site is the source of truth** for all content and commerce; build and seed it only through the Wix connector and the skills below. **Do NOT use the Base44 commerce kit (or any Base44 solution kit).**
-
-The Wix skills live under \`.agents/skills/\` — on ANY turn, read them from that exact path (ignore stray copies like \`agent/skills/\`).
-
-- \`wix-vibe-headless\` — **how the CLIENT is built AND how the site is seeded**: your vertical's UI client + REST scaffolds are already deployed into \`src/\`; one vertical per capability under \`references/<vertical>/\` (storefront, bookings, events, blog, portfolio, restaurants, cms, pricing-plans, members), each with a \`seed/\` module that creates content over the connector.
-- \`wix-docs\` — **fallback** for **frontend code**, **backend code**, or **runtime / API management operations** alike: search + read the Wix API reference docs.
-`;
-const amd = '/app/AGENTS.md';
-const cur = existsSync(amd) ? readFileSync(amd, 'utf8') : '';
-if (!cur.includes('Wix-managed headless frontend')) { appendFileSync(amd, NOTE); deployed.agentsMdPinned = true; }
 
 console.log(JSON.stringify(deployed));
