+++ skills/wix-vibe-headless/platforms/base44.md
@@ -19,8 +19,9 @@ Install two skills — they land under `.agents/skills/`:
   doesn't cover.
 
 Install via the skills CLI — run this through exec_tool, exactly as written. It installs the two
-skills, then runs `deploy.cjs <vertical>`, which lays the shared transport + **your one vertical's**
-REST scaffolds and UI client into `src/` and pins an AGENTS.md note. **Set `VERTICAL`** to what the
+skills, then runs `deploy.cjs <vertical>` (lays the shared transport + **your one vertical's**
+REST scaffolds and UI client into `src/`) and `pin-agents-md.cjs` (pins the project's AGENTS.md
+note so later turns keep the rules). **Set `VERTICAL`** to what the
 prompt asks for — one of `storefront`, `bookings`, `blog`, `cms`, `portfolio`, `pricing-plans`,
 `events`, `members` (too vague to tell? do STEP 2 first, then set it).
 
@@ -37,9 +38,11 @@ for (const skill of ['wix-vibe-headless', 'wix-docs']) {
       : out.includes('No valid skills') ? 'not_found' : 'unknown';
   } catch (e) { results[skill] = 'error: ' + e.message; }
 }
-// Deploy the shared transport + ONLY this vertical's scaffolds/UI into src/, and pin AGENTS.md.
+// Deploy the shared transport + ONLY this vertical's scaffolds/UI into src/.
 const deploy = execSync(`node /app/.agents/skills/wix-vibe-headless/install/deploy.cjs ${VERTICAL}`, { cwd: '/app' }).toString();
-return { results, installed: readdirSync('/app/.agents/skills'), deploy: JSON.parse(deploy) };
+// Pin the project's AGENTS.md note (idempotent) so the rules survive after this doc leaves context.
+const agentsMd = execSync(`node /app/.agents/skills/wix-vibe-headless/install/pin-agents-md.cjs`, { cwd: '/app' }).toString();
+return { results, installed: readdirSync('/app/.agents/skills'), deploy: JSON.parse(deploy), agentsMd: JSON.parse(agentsMd) };
 ```
 
 Read the skills with **`read_file`** (rooted at `/app` → workspace-relative path, e.g.
